# dev_core/api/openapi.yaml
openapi: 3.0.3
info:
  title: InduFrame Control Plane API
  version: 0.1.0
  description: dev_core 面向 dev_ide 的控制面契约。
servers:
  - url: /api/v1
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [user]
      summary: listUsers
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      operationId: createUser
      tags: [user]
      summary: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /users/{userId}:
    put:
      operationId: updateUser
      tags: [user]
      summary: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      operationId: deleteUser
      tags: [user]
      summary: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /projects/{projectId}/runtime-users:
    get:
      operationId: listRuntimeUsers
      tags: [runtimeAccess]
      summary: listRuntimeUsers
      description: 支持 keyword 搜索、page（默认 1）、limit（默认 10，上限 100）。data 返回 list 和 pagination（page、limit、total）。
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      operationId: createRuntimeUser
      tags: [runtimeAccess]
      summary: createRuntimeUser
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeUserCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /projects/{projectId}/runtime-users/{userId}/status:
    patch:
      operationId: updateRuntimeUserStatus
      tags: [runtimeAccess]
      summary: updateRuntimeUserStatus
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeUserStatusRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /projects/{projectId}/runtime-users/{userId}:
    delete:
      operationId: deleteRuntimeUser
      tags: [runtimeAccess]
      summary: deleteRuntimeUser
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /projects/{projectId}/runtime-users/{userId}/roles:
    put:
      operationId: replaceRuntimeUserRoles
      tags: [runtimeAccess]
      summary: replaceRuntimeUserRoles
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceIDsRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /projects/{projectId}/runtime-users/{userId}/reset-password:
    post:
      operationId: resetRuntimeUserPassword
      tags: [runtimeAccess]
      summary: resetRuntimeUserPassword
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /projects/{projectId}/runtime-roles:
    get:
      operationId: listRuntimeRoles
      tags: [runtimeAccess]
      summary: listRuntimeRoles
      description: 支持 keyword 搜索、page（默认 1）、limit（默认 10，上限 100）。data 返回 list 和 pagination（page、limit、total）。
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      operationId: createRuntimeRole
      tags: [runtimeAccess]
      summary: createRuntimeRole
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeRoleCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /projects/{projectId}/runtime-roles/{roleId}:
    put:
      operationId: updateRuntimeRole
      tags: [runtimeAccess]
      summary: updateRuntimeRole
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeRoleUpdateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      operationId: deleteRuntimeRole
      tags: [runtimeAccess]
      summary: deleteRuntimeRole
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: NodeAgentToken
  schemas:
...
    UserCreateRequest:
      type: object
      additionalProperties: false
      description: 仅组织管理员可创建用户。仅用户名、密码、角色必填，其余为选填高级资料。
      required: [username, password, role]
      properties:
        username: { type: string, minLength: 3, maxLength: 50 }
        password: { type: string, format: password, minLength: 1, description: 非空临时密码；首次登录必须改密，新密码至少8个字符且不同于旧密码，不要求字符组合 }
        email: { type: string, maxLength: 254 }
        phone: { type: string, maxLength: 40 }
        fullName: { type: string, maxLength: 100 }
        gender: { type: string, enum: ['', male, female, other, undisclosed] }
        attributes:
          type: object
          description: 自定义文本资料，最多20项；名称不重复且不使用账号或权限字段名，不参与鉴权。
          maxProperties: 20
          additionalProperties: { type: string, maxLength: 500 }
        role: { type: string, enum: [SYSTEM_ADMIN, PROJECT_ADMIN, OPS_ADMIN] }
        status: { type: string, enum: [active, inactive, suspended] }
    UserUpdateRequest:
      type: object
      additionalProperties: false
      description: 用户名不可修改；不传字段保留，空字符串或空属性对象清空。不能修改自己的角色或状态。
      properties:
        email: { type: string, maxLength: 254 }
        phone: { type: string, maxLength: 40 }
        fullName: { type: string, maxLength: 100 }
        gender: { type: string, enum: ['', male, female, other, undisclosed] }
        attributes:
          type: object
          description: 自定义文本资料，最多20项；名称不重复且不使用账号或权限字段名，不参与鉴权。
          maxProperties: 20
          additionalProperties: { type: string, maxLength: 500 }
        role: { type: string, enum: [SYSTEM_ADMIN, PROJECT_ADMIN, OPS_ADMIN] }
        status: { type: string, enum: [active, inactive, suspended] }
...
    ReplaceIDsRequest:
      type: object
      properties:
        ids: { type: array, items: { type: string, format: uuid } }
        roleIds: { type: array, items: { type: string, format: uuid } }
        tagIds: { type: array, items: { type: string, format: uuid } }
...
    RuntimeUserCreateRequest:
      type: object
      required: [username, initialPassword]
      properties:
        username: { type: string }
        initialPassword: { type: string, format: password, minLength: 8 }
        roleIds: { type: array, items: { type: string, format: uuid } }
    RuntimeUserStatusRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [active, disabled] }
    ResetPasswordRequest:
      type: object
      required: [newPassword]
      properties:
        newPassword: { type: string, format: password, minLength: 8 }
    RuntimeRoleCreateRequest:
      type: object
      required: [code, name]
      properties:
        code: { type: string }
        name: { type: string }
        description: { type: string }
        capabilities: { type: array, items: { type: string } }
    RuntimeRoleUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/RuntimeRoleCreateRequest'
...
    ApiResponse:
      type: object
      required: [code, msg, data, reqId]
      properties:
        code:
          type: integer
        msg:
          type: string
        data:
          nullable: true
        reqId:
          type: string
security:
  - bearerAuth: []